WIP: refactor for monopolar model
